+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Cere Stats
 
 ## vNext
-- ...
+- [BE] Add ss58 prefix migration for second priority data
 
 ## v0.26.0
 - [BE] Add ss58 prefix migration for first priority data